Refactored & updated ModelTest to ModelTestCase
