+++ tests/Customer/CustomerTestCase.php
@@ -4,10 +4,10 @@ namespace Grixu\SociusModels\Tests\Customer;
 
 use Grixu\SociusModels\Customer\Factories\CustomerFactory;
 use Grixu\SociusModels\Customer\Models\Customer;
-use Grixu\SociusModels\Tests\Helpers\ModelTest;
+use Grixu\SociusModels\Tests\Helpers\ModelTestCase;
 use Illuminate\Database\Eloquent\Model;
 
-class CustomerTest extends ModelTest
+class CustomerTestCase extends ModelTestCase
 {
     protected string $model = Customer::class;
     protected string $factory = CustomerFactory::class;
